+++ src/renderer/src/components/Menu.vue
@@ -34,7 +34,6 @@
 <script lang="ts" setup>
 import { useMenuStore } from '@/store/menu'
 import { MenuComposite } from '@/types/menu'
-import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 
 const { t } = useI18n({ useScope: 'global' })
