+++ src/renderer/src/components/wavy_items/DecimalBlock.vue
@@ -1,0 +1,6 @@
+<template></template>
+
+<script setup lang="ts">
+</script>
+
+<style lang="scss" scoped></style>
