+++ src/renderer/src/components/wavy_items/StringBlock.vue
@@ -1,0 +1,10 @@
+<template>
+  <v-list-item>
+
+  </v-list-item>
+</template>
+
+<script setup lang="ts">
+</script>
+
+<style lang="scss" scoped></style>
