+++ src/renderer/src/plugins/floating_vue.ts
@@ -1,6 +1,6 @@
+import FloatingVue from 'floating-vue'
 import 'floating-vue/dist/style.css'
 import type { App } from 'vue'
-import FloatingVue from 'floating-vue'
 
 export default {
   install: (app: App) => {
