+++ src/renderer/src/views/ProjectManager.vue
@@ -87,8 +87,7 @@ import { Project, ProjectType } from '@W/types/project'
 import { defaultId } from '@W/util/SnowflakeId'
 import { plainToInstance } from 'class-transformer'
 import { storeToRefs } from 'pinia'
-import { computed } from 'vue'
-import { onDeactivated, onMounted, onUnmounted, ref, watch } from 'vue'
+import { computed, onDeactivated, onMounted, onUnmounted, ref, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
 
 const appStore = useAppStore()
